be/c: introduce platform specific `*.c`-files
