+++ include/shared.c
@@ -1,0 +1,76 @@
+/*
+
+This file is part of the Fuzion language implementation.
+
+The Fuzion language implementation is free software: you can redistribute it
+and/or modify it under the terms of the GNU General Public License as published
+by the Free Software Foundation, version 3 of the License.
+
+The Fuzion language implementation is distributed in the hope that it will be
+useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public
+License for more details.
+
+You should have received a copy of the GNU General Public License along with The
+Fuzion language implementation.  If not, see <https://www.gnu.org/licenses/>.
+
+*/
+
+/*-----------------------------------------------------------------------
+ *
+ * Tokiwa Software GmbH, Germany
+ *
+ * Source of shared code of Fuzion C backend.
+ *
+ *---------------------------------------------------------------------*/
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <stdint.h>
+#include <stdbool.h>
+
+
+/**
+ * Perform bitwise comparison of two float values. This is used by
+ * concur.atmic.compare_and_swap/set to compare floats. In particular, this
+ * results is unequality of +0 and -0 and equality of NaN unless undefined bits
+ * differ, etc.
+ *
+ * @param f1, f2 two float values
+ *
+ * @return true iff f1 and f2 are represented in memory by the same bit patternsx.
+ */
+bool fzE_bitwise_compare_float(float f1, float f2)
+{
+  union
+  {
+    float f;
+    int32_t bits;
+  } v1, v2;
+  v1.f = f1;
+  v2.f = f2;
+  return v1.bits == v2.bits;
+}
+
+
+/**
+ * Perform bitwise comparison of two double values. This is used by
+ * concur.atmic.compare_and_swap/set to compare floats. In particular, this
+ * results is unequality of +0 and -0 and equality of NaN unless undefined bits
+ * differ, etc.
+ *
+ * @param d1, d2 two double values
+ *
+ * @return true iff d1 and d2 are represented in memory by the same bit patterns.
+ */
+bool fzE_bitwise_compare_double(double d1, double d2)
+{
+  union
+  {
+    double d;
+    int64_t bits;
+  } v1, v2;
+  v1.d = d1;
+  v2.d = d2;
+  return v1.bits == v2.bits;
+}
