+++ include/win.c
@@ -1,0 +1,383 @@
+/*
+
+This file is part of the Fuzion language implementation.
+
+The Fuzion language implementation is free software: you can redistribute it
+and/or modify it under the terms of the GNU General Public License as published
+by the Free Software Foundation, version 3 of the License.
+
+The Fuzion language implementation is distributed in the hope that it will be
+useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public
+License for more details.
+
+You should have received a copy of the GNU General Public License along with The
+Fuzion language implementation.  If not, see <https://www.gnu.org/licenses/>.
+
+*/
+
+/*-----------------------------------------------------------------------
+ *
+ * Tokiwa Software GmbH, Germany
+ *
+ * Source of windows code of Fuzion C backend.
+ *
+ *---------------------------------------------------------------------*/
+
+// "For example if you want to use winsock2.h you better make sure
+// WIN32_LEAN_AND_MEAN is always defined because otherwise you will
+// get conflicting declarations between the WinSock versions."
+// https://stackoverflow.com/questions/11040133/what-does-defining-win32-lean-and-mean-exclude-exactly#comment108482188_11040230
+#define WIN32_LEAN_AND_MEAN
+
+
+#include <errno.h>
+#include <stdio.h>
+#include <stdlib.h>     // setenv, unsetenv
+#include <string.h>
+#include <stdint.h>
+#include <stdbool.h>
+
+// NYI remove POSIX imports
+#include <sys/stat.h>   // mkdir
+#include <sys/types.h>  // mkdir
+
+#include <winsock2.h>
+#include <windows.h>
+#include <ws2tcpip.h>
+#include <winbase.h>
+
+
+// make directory, return zero on success
+int fzE_mkdir(const char *pathname){
+  return CreateDirectory(pathname, NULL)
+    ? 0
+    : -1;
+}
+
+
+// set environment variable, return zero on success
+int fzE_setenv(const char *name, const char *value, int overwrite){
+  return -1;
+}
+
+
+// unset environment variable, return zero on success
+int fzE_unsetenv(const char *name){
+  return -1;
+}
+
+
+// 0 = blocking
+// 1 = none_blocking
+int fzE_set_blocking(int sockfd, int blocking)
+{
+  u_long b = blocking;
+  return ioctlsocket(sockfd, FIONBIO, &b);
+}
+
+
+// helper function to retrieve
+// the last error that occurred.
+int fzE_net_error()
+{
+  return WSAGetLastError();
+}
+
+
+// fuzion family number -> system family number
+int get_family(int family)
+{
+  return family == 1
+    ? AF_UNIX
+    : family == 2
+    ? AF_INET
+    : family == 10
+    ? AF_INET6
+    : -1;
+}
+
+
+// fuzion socket type number -> system socket type number
+int get_socket_type(int socktype)
+{
+  return socktype == 1
+    ? SOCK_STREAM
+    : socktype == 2
+    ? SOCK_DGRAM
+    : socktype == 3
+    ? SOCK_RAW
+    : -1;
+}
+
+
+// fuzion protocol number -> system protocol number
+int get_protocol(int protocol)
+{
+  return protocol == 6
+    ? IPPROTO_TCP
+    : protocol == 17
+    ? IPPROTO_UDP
+    : protocol == 0
+    ? IPPROTO_IP
+    : protocol == 41
+    ? IPPROTO_IPV6
+    : -1;
+}
+
+
+// close a socket descriptor
+int fzE_close(int sockfd)
+{
+  closesocket(sockfd);
+  return fzE_net_error();
+}
+
+
+// initialize a new socket for given
+// family, socket_type, protocol
+int fzE_socket(int family, int type, int protocol){
+  WSADATA wsaData;
+  return WSAStartup(MAKEWORD(2,2), &wsaData) != 0
+    ? -1
+    : socket(get_family(family), get_socket_type(type), get_protocol(protocol));
+}
+
+
+// get addrinfo structure used for binding/connection of a socket.
+int fzE_getaddrinfo(int family, int socktype, int protocol, int flags, char * host, char * port, struct addrinfo ** result){
+  struct addrinfo hints;
+
+  ZeroMemory(&hints, sizeof(hints));
+
+  hints.ai_family = get_family(family);
+  hints.ai_socktype = get_socket_type(socktype);
+  hints.ai_protocol = get_protocol(protocol);
+  hints.ai_flags = flags;
+
+  return getaddrinfo(host, port, &hints, result);
+}
+
+
+// create a new socket and bind to given host:port
+// result[0] contains either an errorcode or a socket descriptor
+// -1 error, 0 success
+int fzE_bind(int family, int socktype, int protocol, char * host, char * port, int64_t * result){
+  result[0] = fzE_socket(family, socktype, protocol);
+  if (result[0] == -1)
+  {
+    result[0] = fzE_net_error();
+    return -1;
+  }
+  struct addrinfo *addr_info = NULL;
+  int addrRes = fzE_getaddrinfo(family, socktype, protocol, AI_PASSIVE, host, port, &addr_info);
+  if (addrRes != 0)
+  {
+    fzE_close(result[0]);
+    result[0] = addrRes;
+    return -1;
+  }
+  int bind_res = bind(result[0], addr_info->ai_addr, (int)addr_info->ai_addrlen);
+
+  if(bind_res == -1)
+  {
+    fzE_close(result[0]);
+    result[0] = fzE_net_error();
+    return -1;
+  }
+  freeaddrinfo(addr_info);
+  return bind_res;
+}
+
+
+// set the given socket to listening
+// backlog = queuelength of pending connections
+int fzE_listen(int sockfd, int backlog){
+  return ( listen(sockfd, backlog) == -1 )
+    ? fzE_net_error()
+    : 0;
+}
+
+
+// accept a new connection
+// blocks if socket is blocking
+int fzE_accept(int sockfd){
+  return accept(sockfd, NULL, NULL);
+}
+
+
+// create connection for given parameters
+// result[0] contains either an errorcode or a socket descriptor
+// -1 error, 0 success
+int fzE_connect(int family, int socktype, int protocol, char * host, char * port, int64_t * result){
+  // get socket
+  result[0] = fzE_socket(family, socktype, protocol);
+  if (result[0] == -1)
+  {
+    result[0] = fzE_net_error();
+    return -1;
+  }
+  struct addrinfo *addr_info = NULL;
+  int addrRes = fzE_getaddrinfo(family, socktype, protocol, 0, host, port, &addr_info);
+  if (addrRes != 0)
+  {
+    fzE_close(result[0]);
+    result[0] = addrRes;
+    return -1;
+  }
+  int con_res = connect(result[0], addr_info->ai_addr, addr_info->ai_addrlen);
+  if(con_res == -1)
+  {
+    // NYI do we want to try another address in addr_info->ai_next?
+    fzE_close(result[0]);
+    result[0] = fzE_net_error();
+  }
+  freeaddrinfo(addr_info);
+  return con_res;
+}
+
+
+// get the peer's ip address
+// result is the length of the ip address written to buf
+// might return useless information when called on udp socket
+int fzE_get_peer_address(int sockfd, void * buf) {
+  struct sockaddr_storage peeraddr;
+  socklen_t peeraddrlen = sizeof(peeraddr);
+  int res = getpeername(sockfd, (struct sockaddr *)&peeraddr, &peeraddrlen);
+  if (peeraddr.ss_family == AF_INET) {
+    memcpy(buf, &(((struct sockaddr_in *)&peeraddr)->sin_addr.s_addr), 4);
+    return 4;
+  } else if (peeraddr.ss_family == AF_INET6) {
+    memcpy(buf, &(((struct sockaddr_in6 *)&peeraddr)->sin6_addr.s6_addr), 16);
+    return 16;
+  } else {
+    return -1;
+  }
+  return -1;
+}
+
+
+// get the peer's port
+// result is the port number
+// might return useless infomrmation when called on udp socket
+unsigned short fzE_get_peer_port(int sockfd) {
+  struct sockaddr_storage peeraddr;
+  socklen_t peeraddrlen = sizeof(peeraddr);
+  int res = getpeername(sockfd, (struct sockaddr *)&peeraddr, &peeraddrlen);
+  if (peeraddr.ss_family == AF_INET) {
+    return ntohs(((struct sockaddr_in *)&peeraddr)->sin_port);
+  } else if (peeraddr.ss_family == AF_INET6) {
+    return ntohs(((struct sockaddr_in6 *)&peeraddr)->sin6_port);
+  } else {
+    return 0;
+  }
+}
+
+
+// read up to count bytes bytes from sockfd
+// into buf. may block if socket is  set to blocking.
+// return -1 on error or number of bytes read
+int fzE_read(int sockfd, void * buf, size_t count){
+  int rec_res = recvfrom( sockfd, buf, count, 0, NULL, NULL );
+  if (rec_res == -1)
+  {
+    // silently discard rest to
+    // match behaviour on linux
+    return fzE_net_error() == WSAEMSGSIZE
+      ? count
+      : rec_res;
+  }
+  return rec_res;
+}
+
+
+// write buf to sockfd
+// may block if socket is set to blocking.
+// return error code or zero on success
+int fzE_write(int sockfd, const void * buf, size_t count){
+return ( sendto( sockfd, buf, count, 0, NULL, 0 ) == -1 )
+  ? fzE_net_error()
+  : 0;
+}
+
+
+// for 64-bit offset returns the 32 highest bits as a DWORD
+DWORD high_word(off_t value) {
+  return sizeof(off_t) == 4
+    ? 0
+    : (DWORD)(value >> 32);
+}
+
+
+// for 64-bit offset returns the 32 lowest bits as a DWORD
+DWORD low_word(off_t value) {
+  return (DWORD)(value & ((1ULL << 32) - 1));
+}
+
+
+// returns -1 on error, size of file in bytes otherwise
+long fzE_get_file_size(FILE* file) {
+  // store current pos
+  long cur_pos = ftell(file);
+  if(cur_pos == -1 || fseek(file, 0, SEEK_END) == -1){
+    return -1;
+  }
+
+  long size = ftell(file);
+
+  // reset seek position
+  fseek(file, cur_pos, SEEK_SET);
+
+  return size;
+}
+
+
+/*
+ * create a memory map of a file at an offset.
+ * unix:    the offset must be a multiple of the page size, usually 4096 bytes.
+ * windows: the offset must be a multiple of the memory allocation granularity, usually 65536 bytes
+ *          see also, https://devblogs.microsoft.com/oldnewthing/20031008-00/?p=42223
+ *
+ * returns:
+ *   - error   :  result[0]=-1 and NULL
+ *   - success :  result[0]=0  and an address where the file was mapped to
+ */
+void * fzE_mmap(FILE * file, off_t offset, size_t size, int * result) {
+
+  if (fzE_get_file_size(file) < (offset + size)){
+    result[0] = -1;
+    return NULL;
+  }
+
+  HANDLE file_handle = (HANDLE)_get_osfhandle(fileno(file));
+
+  /* "If dwMaximumSizeLow and dwMaximumSizeHigh are 0 (zero), the maximum size of the file mapping
+      object is equal to the current size of the file that hFile identifies.
+      An attempt to map a file with a length of 0 (zero) fails with an error code
+      of ERROR_FILE_INVALID. Applications should test for files with a length of 0 (zero) and reject those files."
+  */
+  HANDLE file_mapping_handle = CreateFileMapping(file_handle, NULL, PAGE_READWRITE, 0, 0, NULL);
+  if (file_mapping_handle == NULL) {
+    result[0] = -1;
+    return NULL;
+  }
+
+  void * mapped_address = MapViewOfFile(file_mapping_handle, FILE_MAP_ALL_ACCESS, high_word(offset), low_word(offset), size);
+  if (mapped_address == NULL) {
+    CloseHandle(file_mapping_handle);
+    result[0] = -1;
+    return NULL;
+  }
+  result[0] = 0;
+  return mapped_address;
+}
+
+
+// unmap an address that was previously mapped by fzE_mmap
+// -1 error, 0 success
+int fzE_munmap(void * mapped_address, const int file_size){
+  return UnmapViewOfFile(mapped_address)
+    ? 0
+    : -1;
+}
+
